+++ operations.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void add(matrix *m, int to, int what, double scalar, char orient);
-void mul(matrix *m, int what, double scalar, char orient);
+void add(matrix *m, int to, int what,long double scalar, char orient);
+void mul(matrix *m, int what,long double scalar, char orient);
 void swap(matrix *m, int what, int with, char orient);
